TimestampParser naming refactoring
