+++ embulk-core/src/main/java/org/embulk/spi/time/TimestampParser.java
@@ -6,9 +6,10 @@ import org.embulk.config.Task;
 import org.embulk.config.Config;
 import org.embulk.config.ConfigInject;
 import org.embulk.config.ConfigDefault;
+import org.jruby.Ruby;
 import org.jruby.util.RubyDateFormatter;
 import org.jruby.util.RubyDateParser;
-import org.jruby.util.Temporal;
+import org.jruby.util.RubyDateParser.Temporal;
 
 import java.util.List;
 
@@ -41,14 +42,15 @@ public class TimestampParser
     private TimestampParser(ScriptingContainer jruby, String format, DateTimeZone defaultTimeZone)
     {
         // TODO get default current time from ExecTask.getExecTimestamp
-        this.parser = new RubyDateParser(jruby.getProvider().getRuntime().getCurrentContext());
-        this.compiledPattern = this.parser.compilePattern(format);
+        Ruby runtime = jruby.getProvider().getRuntime();
+        this.parser = new RubyDateParser(runtime.getCurrentContext());
+        this.compiledPattern = this.parser.compilePattern(runtime.newString(format), true);
         this.defaultTimeZone = defaultTimeZone;
     }
 
     public Timestamp parse(String text) throws TimestampParseException
     {
-        Temporal tmp = parser.date_strptime(compiledPattern, text);
+        Temporal tmp = parser.parse(compiledPattern, text);
         String zone = tmp.getZone();
         DateTimeZone timeZone = defaultTimeZone;
         if (zone != null) {
