+++ embulk-jruby-extension/test/mri/date/test_date_strptime_simple.rb
